Move the `agent` cog's Config, Input, and Output classes to separate files
