+++ lib/roast/dsl/cogs/agent/input.rb
@@ -1,0 +1,43 @@
+# typed: true
+# frozen_string_literal: true
+
+module Roast
+  module DSL
+    module Cogs
+      class Agent < Cog
+        class Input < Cog::Input
+          #: String?
+          attr_accessor :prompt
+
+          #: String?
+          attr_accessor :session
+
+          #: () -> void
+          def initialize
+            super
+            @prompt = nil #: String?
+          end
+
+          #: () -> void
+          def validate!
+            valid_prompt!
+          end
+
+          #: (untyped) -> void
+          def coerce(input_return_value)
+            if input_return_value.is_a?(String)
+              self.prompt ||= input_return_value
+            end
+          end
+
+          #: () -> String
+          def valid_prompt!
+            raise Cog::Input::InvalidInputError, "'prompt' is required" unless @prompt.present?
+
+            @prompt
+          end
+        end
+      end
+    end
+  end
+end
